+++ src/modules/stock-logs/models/distribution-house-stock-log.model.ts
@@ -21,18 +21,18 @@ import { Transport } from '../../transports/models/transport.model';
 export class DistributionHouseStockLog extends Model {
   @PrimaryKey
   @AutoIncrement
-  @Column(DataType.INTEGER)
+  @Column(DataType.BIGINT)
   declare id: number;
 
   @ForeignKey(() => DistributionHouse)
-  @Column(DataType.INTEGER)
+  @Column(DataType.BIGINT)
   declare distributionHouseId: number;
 
   @Column(DataType.DATE)
   declare date: Date;
 
   @ForeignKey(() => Product)
-  @Column(DataType.INTEGER)
+  @Column(DataType.BIGINT)
   declare productId: number;
 
   @Column({
@@ -62,17 +62,17 @@ export class DistributionHouseStockLog extends Model {
   @Column(DataType.STRING(50))
   declare referenceType: string;
 
-  @Column(DataType.INTEGER)
+  @Column(DataType.BIGINT)
   declare referenceId: number;
 
   @ForeignKey(() => Transport)
-  @Column(DataType.INTEGER)
+  @Column(DataType.BIGINT)
   declare transportId: number;
 
   @Column(DataType.TEXT)
   declare notes: string;
 
-  @Column(DataType.INTEGER)
+  @Column(DataType.BIGINT)
   declare createdBy: number;
 
   @CreatedAt
